Cloud Config certs as a dynamic template
We want to avoid using hardcoded values in the user_data
cloud config, to avoid duplication and in to use some
kind of automated certificate generation in the future.

We also move the configuration to a separate file, using
template_file. Using template_file would allow us to reuse
the value and print it for debugging.